Use dash controlled text tracks
As of video.js v5.14, video.js no longer supports native
subtitles on anything besides Safari. For Chrome, this means
there is *no* subtitle support with mpeg-dash. This will wait for
dash.js to load all the text tracks and then add them to video.js
if we're using a version that does not attempt to use native
subtitles. When the video.js menu is used to change captions,
we'll set the corresponding active text track in dash.

We *cannot* dynamically add text track cues to video.js beacuse
fragmented text tracks are loaded on the fly and there's no event
fired when they are loaded. Therefore, we must has dash to
natively display captions.