Load text tracks from dash into video.js (#135)
* Use dash controlled text tracks

As of video.js v5.14, video.js no longer supports native
subtitles on anything besides Safari. For Chrome, this means
there is *no* subtitle support with mpeg-dash. This will wait for
dash.js to load all the text tracks and then add them to video.js
if we're using a version that does not attempt to use native
subtitles. When the video.js menu is used to change captions,
we'll set the corresponding active text track in dash.

We *cannot* dynamically add text track cues to video.js beacuse
fragmented text tracks are loaded on the fly and there's no event
fired when they are loaded. Therefore, we must has dash to
natively display captions.

* Update README.md to describe `nativeCaptions` usage requirements

* Require videojs@v5.18

The v5.18.0 release of video.js includes a fix for https://github.com/videojs/video.js/issues/4093 with https://github.com/videojs/video.js/pull/4131. This is neccessary for text tracks to work in Chrome.

* fixup! Use dash controlled text tracks

- Only set `textTrack.mode` for `captions` and `subtitles.
- Don't call the initial `updateActiveDashTextTrack` until after
  the `textTrack.mode`s have been set.